Cleanup BoardLoader.
